念の為ファイルの存在確認

+++ meson/install_macos.py
@@ -1,8 +1,12 @@
+import os
 import shutil
 import subprocess
 import sys
 
 def install(src, dst):
+	if not os.path.isfile(src):
+		print('error: ' + src + ' not found.')
+		exit(1)
 	print('Installing ' + src + ' to ' + dst)
 	shutil.copy(src, dst)
